mobilitaet_dtv/pyproject.toml
[project]
name = "mobilitaet-dtv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "filehash>=0.2.dev1",
    "more-itertools>=10.6.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
